Former-commit-id: f61c07530d612d3f6ac97691c3efb34d284eb24f

+++ source/display.hpp
@@ -83,7 +83,7 @@ namespace baal
 		
 		int run(double _runtime, float _timestep)
         {
-        	int layer = network.getNeuronPopulations().size() - 1;
+        	uint64_t layer = network.getNeuronPopulations().size() - 1;
         	engine->rootContext()->setContextProperty("layers", layer);
             std::thread spikeManager([this, _runtime, _timestep]{
                 network.run(_runtime, _timestep);
